Failed attempt at a progress bar
My attempt at a progress bar failed because the server doesn't respond to the ajax request until the main process finishes.
However I used some of the javascript code to force a reload when the page hangs.
The hang has me totally perplexed; it happens at random, and possibly only on Firefox.